requirements.txt
aiohttp==3.10.11
PyJWT==2.9.0
